extract delete settings into separate component
